% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{blur_point}
\alias{blur_point}
\title{blur ellipses function}
\usage{
blur_point(
  x,
  levels = seq(0.001, 1 - 0.1, l = 15),
  alpha_mult,
  col = "black",
  center
)
}
\arguments{
\item{x}{An object. In the default method the parameter x should be a correlation between -1 and 1 or a square positive definite matrix at least 2x2 in size. It will be treated as the correlation or covariance of a multivariate normal distribution.}

\item{levels}{contour levels}

\item{alpha_mult}{multiplier on transparency level}

\item{col}{default is black}

\item{center}{two-vector giving center of ellipse}
}
\description{
blur ellipses function
}
