% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overwrite.R
\name{check_overwrite}
\alias{check_overwrite}
\title{Check overwrite}
\usage{
check_overwrite(method, return.paths, ...)
}
\arguments{
\item{method}{passed from animate_paths()}

\item{return.paths}{passed from animate_paths()}

\item{...}{passed from animate_paths(); used to check for user-specified value for img.name}
}
\value{
NULL, unless there is risk of overwritting and the user interrupts animation (\code{FALSE})
}
\description{
Check overwrite
}
