% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_interpolation.R
\name{covariate_interp}
\alias{covariate_interp}
\title{Synchronous interpolation of covariate using either GAM (same as paths) or piece-wise constant if covariate is a factor}
\usage{
covariate_interp(paths, covariate = NULL, Time.name, time.grid, s_args)
}
\arguments{
\item{paths}{lists of data.frames containing positions, times, and covariate for each individual}

\item{covariate}{character string giving name of covariate variable in data.frames}

\item{Time.name}{character string giving name of time variable in data.frames}

\item{time.grid}{grid of possible times to use for interpolation (individuals will only be interpolated to times within the range of observation times)}

\item{s_args}{arguments to \code{mgcv::s()} for GAM interpolation method}
}
\value{
list of interpolated covariate by individual
}
\description{
Synchronous interpolation of covariate using either GAM (same as paths) or piece-wise constant if covariate is a factor
}
