% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_paths.R
\name{get.network.colors}
\alias{get.network.colors}
\title{get.network.colors()
Finds all maximal cliques in the network at each time point and tries to assign them a useful coloring}
\usage{
get.network.colors(binary.network, network.color.options = NULL)
}
\arguments{
\item{binary.network}{a 3D array giving the time-varying adjecency matrix of a dynamic network.}

\item{network.color.options}{vector of colors}
}
\value{
a list of two elements: a list of the maximal cliques at each time, and c list with colors for each clique at each time
}
\description{
get.network.colors()
Finds all maximal cliques in the network at each time point and tries to assign them a useful coloring
}
