% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_googlemap_min_scale}
\alias{get_googlemap_min_scale}
\title{Figure out scale and centering of google map by transforming reported lat long bounding box back to web mercator}
\usage{
get_googlemap_min_scale(map)
}
\arguments{
\item{map}{\code{ggmap} object}
}
\value{
scale (factor by which web mercator has been shrunk) and min (leftmost, bottom most coordinate of rectangle)
}
\description{
Figure out scale and centering of google map by transforming reported lat long bounding box back to web mercator
}
