% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{googlemap_proj}
\alias{googlemap_proj}
\title{adjust center + scale for google map plotting}
\usage{
googlemap_proj(x, map)
}
\arguments{
\item{x}{\code{sf} object}

\item{map}{\code{ggmap} object}
}
\value{
two-column matrix of locations from \code{x} projected to match \code{map}
}
\description{
adjust center + scale for google map plotting
}
