% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_colors.R
\name{network_interp}
\alias{network_interp}
\title{Synchronous interpolation of network using piece-wise constant interpolation}
\usage{
network_interp(network = NULL, network.times, time.grid)
}
\arguments{
\item{network}{array of network observations of dimension (\code{n.indiv, n.indiv, length(network.times)})}

\item{network.times}{vector of times at which network observations are made}

\item{time.grid}{times at which network will be interpolated}
}
\value{
array of dimension \code{n.indiv, n.indiv, length(time.grid))}
}
\description{
Synchronous interpolation of network using piece-wise constant interpolation
}
