% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vultures}
\alias{vultures}
\title{GPS locations of turkey vultures.}
\format{
A data frame with 215719 rows and 11 variables:
\describe{
  \item{timestamp}{time of observation}
  \item{location.long}{logitude}
  \item{location.lat}{latitude}
  \item{individual.local.identifier}{identifier for each individual}
  ...
}
}
\source{
\url{https://www.datarepository.movebank.org/handle/10255/move.362/} Bildstein K, Barber D, Bechard MJ (2014) Data from: Environmental drivers of variability in the movement ecology of turkey vultures (Cathartes aura) in North and South America. Movebank Data Repository. doi:10.5441/001/1.46ft1k05
}
\usage{
vultures
}
\description{
A dataset containing a subset of the locations of turkey vultures (2003--2006), with time stamps, from:
}
\details{
Dodge S, Bohrer G, Bildstein K, Davidson SC, Weinzierl R, Mechard MJ, Barber D, Kays R, Brandes D, Han J (2014) Environmental drivers of variability in the movement ecology of turkey vultures (Cathartes aura) in North and South America. Philosophical Transactions of the Royal Society B 20130195. doi:10.1098/rstb.2013.0195

  Bildstein K, Barber D, Bechard MJ (2014) Data from: Environmental drivers of variability in the movement ecology of turkey vultures (Cathartes aura) in North and South America. Movebank Data Repository. doi:10.5441/001/1.46ft1k05
}
\keyword{datasets}
