% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whales}
\alias{whales}
\title{GPS locations of three species of whales.}
\format{
A data frame with 4303 rows and 4 variables:
\describe{
  \item{timestamp}{time of observation}
  \item{location.long}{logitude}
  \item{location.lat}{latitude}
  \item{individual.local.identifier}{identifier for each individual}
  ...
}
}
\source{
\url{doi:10.5441/001/1.98f5r6d0} Irvine LM, Follett TM, Winsor MH, Mate BR, Palacios DM (2020) Data from: Study "Blue and fin whales Southern California 2014-2015 - Argos data". Movebank Data Repository.
}
\usage{
whales
}
\description{
A dataset containing locations of whales, with time stamps, from:
}
\details{
Irvine LM, Winsor MH, Follett TM, Mate BR, Palacios DM (2020) An at-sea assessment of Argos location accuracy for three species of large whales, and the effect of deep-diving behavior on location error. Animal Biotelemetry 8:20.
 
 Irvine LM, Follett TM, Winsor MH, Mate BR, Palacios DM (2020) Data from: Study "Blue and fin whales Southern California 2014-2015 - Argos data". Movebank Data Repository. doi:10.5441/001/1.98f5r6d0
}
\keyword{datasets}
